Moves common functions to utils
